+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the apache2 cookbook.
 
 ## Unreleased
 
+- Add `envvars_additional_params` property to install resource
+
 ## 8.13.1 - *2021-08-30*
 
 - Standardise files with files in sous-chefs/repo-management